+++ RCAIDE/Library/Methods/Noise/Common/__init__.py
@@ -18,3 +18,4 @@ from .generate_zero_elevation_microphone_locations       import generate_zero_el
 from .generate_terrain_microphone_locations              import generate_terrain_microphone_locations
 from .generate_hemisphere_microphone_locations           import generate_hemisphere_microphone_locations
 from .compute_relative_noise_evaluation_locations        import compute_relative_noise_evaluation_locations 
+from .post_process_noise_data                            import post_process_noise_data
