+++ RCAIDE/Library/Methods/Noise/Correlation_Buildup/Turbofan/turbofan_engine_noise.py
@@ -107,9 +107,9 @@ def turbofan_engine_noise(microphone_locations, turbofan, aeroacoustic_data, seg
     RCAIDE.Library.Methods.Noise.Common.SPL_arithmetic
     """
     # unpack   
-    Velocity_primary        = turbofan.core_nozzle.exit_velocity * np.ones_like(aeroacoustic_data.core_nozzle.exit_velocity)  # aeroacoustic_data.core_nozzle.exit_velocity or use mass flow rate 
-    Velocity_secondary      = turbofan.fan_nozzle.exit_velocity * np.ones_like(aeroacoustic_data.core_nozzle.exit_velocity)  # aeroacoustic_data.fan_nozzle.exit_velocity or use mass flow rate 
-    N1                      = aeroacoustic_data.low_pressure_spool.angular_velocity / Units.rpm
+    Velocity_primary        = turbofan.core_nozzle.noise_speed * np.ones_like(aeroacoustic_data.core_nozzle.exit_velocity)  # aeroacoustic_data.core_nozzle.exit_velocity or use mass flow rate 
+    Velocity_secondary      = turbofan.fan_nozzle.noise_speed * np.ones_like(aeroacoustic_data.core_nozzle.exit_velocity)  # aeroacoustic_data.fan_nozzle.exit_velocity or use mass flow rate 
+    N1                      = turbofan.fan.angular_velocity* np.ones_like(aeroacoustic_data.core_nozzle.exit_velocity)  / Units.rpm
 
     Temperature_secondary  = aeroacoustic_data.fan_nozzle.exit_stagnation_temperature[:,0] 
     Pressure_secondary     = aeroacoustic_data.fan_nozzle.exit_stagnation_pressure[:,0] 
@@ -122,8 +122,8 @@ def turbofan_engine_noise(microphone_locations, turbofan, aeroacoustic_data, seg
     distance_microphone    = np.linalg.norm(microphone_locations,axis = 1)    
     Diameter_primary       = turbofan.core_nozzle.diameter
     Diameter_secondary     = turbofan.fan_nozzle.diameter
-    engine_height          = turbofan.height
-    EXA                    = turbofan.length /  turbofan.diameter 
+    engine_height          = turbofan.origin[0][2]
+    EXA                    = turbofan.engine_length /  turbofan.engine_diameter 
     Plug_diameter          = turbofan.plug_diameter 
     Xe                     = turbofan.geometry_xe
     Ye                     = turbofan.geometry_ye
